Fix sha1 function
From my notes:

haskell-gsasl in the sha1 (and probably the other hash) function expects that
gsasl_sha1 takes in a pointer to an allocated 20-byte buffer and writes into
that.  In fact, gsasl_sha1 takes in a pointer and allocates and writes into it
a 20-bytes buffer that the caller is expected to free.  The current
implementation results in hashes of all 0 coming back.